+++ docs/history.rst
@@ -5,3 +5,5 @@ Release History
 
 Next Release
 ------------
+
+- Import @briank's work on prepit.